intermediate commit; refactor to util
