+++ src/dataprob/model_wrapper/model_wrapper.py
@@ -6,11 +6,11 @@ from dataprob.model_wrapper._function_processing import analyze_fcn_sig
 from dataprob.model_wrapper._function_processing import reconcile_fittable
 from dataprob.model_wrapper._function_processing import param_sanity_check
 
-from dataprob.model_wrapper._dataframe_processing import read_spreadsheet
+from dataprob.util.read_spreadsheet import read_spreadsheet
 from dataprob.model_wrapper._dataframe_processing import validate_dataframe
 from dataprob.model_wrapper._dataframe_processing import param_into_existing
 
-from dataprob.check import check_float
+from dataprob.util.check import check_float
 
 import numpy as np
 import pandas as pd
