+++ src/dataprob/plot/plot_residuals.py
@@ -6,7 +6,7 @@ from dataprob.plot._plot_utils import get_plot_features
 from dataprob.plot._plot_utils import get_style
 from dataprob.plot._plot_utils import get_vectors
 from dataprob.plot._plot_utils import get_plot_dimensions
-from dataprob.check import check_bool
+from dataprob.util.check import check_bool
 
 import matplotlib
 from matplotlib import pyplot as plt
