+++ src/dataprob/util/check.py
